@@ -18,7 +18,7 @@ void main() {
 }
 
 class ExampleApp extends StatefulWidget {
-  const ExampleApp({Key? key}) : super(key: key);
+  const ExampleApp({super.key});
 
   @override
   _ExampleAppState createState() => _ExampleAppState();
