@@ -30,7 +30,7 @@ const _asset2 = 'nasa_on_a_mission.mp3';
 class SourcesTab extends StatefulWidget {
   final AudioPlayer player;
 
-  const SourcesTab({Key? key, required this.player}) : super(key: key);
+  const SourcesTab({super.key, required this.player});
 
   @override
   State<SourcesTab> createState() => _SourcesTabState();
